LayoutTests/imported/w3c/web-platform-tests/navigation-api/currententrychange-event/constructor-expected.txt


PASS can't bypass required members by omitting the dictionary entirely
PASS from is required
FAIL all properties are reflected back Type error
FAIL defaults are as expected Type error
